Retrieve date and username per changelist number.
It is not possible to use a single invocation of p4 filelog to get
revision data for all changelists returned by p4 annotate.  To prevent
invoking p4 filelog twice (once with -h and once without), retrieve
data for each changelist number individually.

To improve performance, changelists are stored in a map for later reuse.